feat: rename e2e test to reflect their purpose
